+++ noodles-bam/src/record/convert.rs
@@ -1,6 +1,6 @@
 use std::{io, str};
 
-use noodles_sam::{self as sam, RecordExt};
+use noodles_sam::{self as sam, AlignmentRecord};
 
 use super::{Record, ReferenceSequenceId};