+++ noodles-cram/src/data_container/compression_header/preservation_map/substitution_matrix/builder.rs
@@ -1,4 +1,4 @@
-use noodles_sam::RecordExt;
+use noodles_sam::AlignmentRecord;
 
 use crate::{record::Feature, Record};